added welcome subpages
